requirements.txt
streamlit==1.39.0
mysql-connector-python==9.1.0
pandas==2.2.3
plotly==5.24.1
